fixed bug with datatype in transaction

+++ src/main/java/io/swagger/model/Transaction.java
@@ -212,7 +212,7 @@ public class Transaction   {
     return performerID;
   }
 
-  public void setPerformerID(Integer performerID) {
+  public void setPerformerID(Long performerID) {
     this.performerID = performerID;
   }
